Re-Arrange file structure
