+++ .gitignore
@@ -1,1 +1,2 @@
 /venv/Lib/site-packages/
+/Classes/gibberish.py
