+++ Classes/Settings.py
@@ -1,0 +1,1 @@
+countryformat = 'de'
